(PC-17806) refactor(auth): move login screen in pages folder
