+++ src/features/auth/signup/AfterSignupEmailValidationBuffer/AfterSignupEmailValidationBuffer.tsx
@@ -4,7 +4,7 @@ import React, { useEffect } from 'react'
 import { api } from 'api/api'
 import { AccountState, ValidateEmailResponse } from 'api/gen'
 import { useValidateEmailMutation } from 'features/auth/api/useValidateEmailMutation'
-import { useLoginRoutine } from 'features/auth/login/useLoginRoutine'
+import { useLoginRoutine } from 'features/auth/helpers/useLoginRoutine'
 import { UseNavigationType, UseRouteType } from 'features/navigation/RootNavigator/types'
 import { homeNavConfig } from 'features/navigation/TabBar/helpers'
 import { isTimestampExpired } from 'libs/dates'
